Op.11 No.10: proofreading
The tempo marking does not follow either SrcA or SrcB. However, SrcA and
SrcB only agree on the textual tempo mark, not the metronome mark.

Follow the SrcB marking. See the inline comment for rationale.

(Note that I opted to only adjust the textual tempo marking at this
time, not any \tempo indications.)